refactor(core): fix test
